added parallelization as a bonus for pulling my merge request :)
