+++ cea/thermal_loads.py
@@ -12,9 +12,12 @@ from __future__ import division
 
 import numpy as np
 import pandas as pd
+from geopandas import GeoDataFrame
+
 import contributions.thermal_loads_new_ventilation.ventilation
 import hvac_kaempf
 import functions
+from contributions.thermal_loads_new_ventilation import simple_window_generator as simple_window_generator
 
 
 def calc_tHC_corr(SystemH, SystemC, sys_e_ctrl):
@@ -881,3 +884,477 @@ def Calc_Rf_sh (ShadingType):
 
 if __name__ == '__main__':
     test_thermal_loads_new_ventilation()
+
+
+class BuildingProperties(object):
+    """
+    Groups building properties used for the calc-thermal-loads functions. Stores the full DataFrame for each of the
+    building properties and provides methods for indexing them by name.
+
+    G. Happle   BuildingPropsThermalLoads   27.05.2016
+    """
+
+    def __init__(self, locator, gv):
+        """
+        Read building properties from input shape files and construct a new BuildingProperties object.
+
+        PARAMETERS
+        ----------
+
+        :param locator: an InputLocator for locating the input files
+        :type locator: cea.inputlocator.InputLocator
+
+        :param gv: contains the context (constants and models) for the calculation
+        :type gv: cea.globalvar.GlobalVariables
+
+        RETURNS
+        -------
+
+        :returns: object of type BuildingProperties
+        :rtype: BuildingProperties
+
+        INPUT / OUTPUT FILES
+        --------------------
+
+        - get_radiation: C:\reference-case\baseline\outputs\data\solar-radiation\radiation.csv
+        - get_surface_properties: C:\reference-case\baseline\outputs\data\solar-radiation\properties_surfaces.csv
+        - get_building_geometry: C:\reference-case\baseline\inputs\building-geometry\zone.shp
+        - get_building_hvac: C:\reference-case\baseline\inputs\building-properties\technical_systems.shp
+        - get_building_thermal: C:\reference-case\baseline\inputs\building-properties\thermal_properties.shp
+        - get_building_occupancy: C:\reference-case\baseline\inputs\building-properties\occupancy.shp
+        - get_building_architecture: C:\reference-case\baseline\inputs\building-properties\architecture.shp
+        - get_building_age: C:\reference-case\baseline\inputs\building-properties\age.shp
+        - get_building_comfort: C:\reference-case\baseline\inputs\building-properties\indoor_comfort.shp
+        - get_building_internal: C:\reference-case\baseline\inputs\building-properties\internal_loads.shp
+        """
+        self.gv = gv
+        gv.log("reading input files")
+        solar = pd.read_csv(locator.get_radiation()).set_index('Name')
+        surface_properties = pd.read_csv(locator.get_surface_properties())
+        prop_geometry = GeoDataFrame.from_file(locator.get_building_geometry())
+        prop_geometry['footprint'] = prop_geometry.area
+        prop_geometry['perimeter'] = prop_geometry.length
+        prop_geometry = prop_geometry.drop('geometry', axis=1).set_index('Name')
+        prop_hvac = GeoDataFrame.from_file(locator.get_building_hvac()).drop('geometry', axis=1)
+        prop_thermal = GeoDataFrame.from_file(locator.get_building_thermal()).drop('geometry', axis=1).set_index('Name')
+        prop_occupancy_df = GeoDataFrame.from_file(locator.get_building_occupancy()).drop('geometry', axis=1).set_index(
+            'Name')
+        prop_occupancy = prop_occupancy_df.loc[:, (prop_occupancy_df != 0).any(
+            axis=0)]  # trick to erase occupancies that are not being used (it speeds up the code)
+        prop_architecture = GeoDataFrame.from_file(locator.get_building_architecture()).drop('geometry',
+                                                                                             axis=1).set_index('Name')
+        prop_age = GeoDataFrame.from_file(locator.get_building_age()).drop('geometry', axis=1).set_index('Name')
+        prop_comfort = GeoDataFrame.from_file(locator.get_building_comfort()).drop('geometry', axis=1).set_index('Name')
+        prop_internal_loads = GeoDataFrame.from_file(locator.get_building_internal()).drop('geometry',
+                                                                                           axis=1).set_index('Name')
+        # get temperatures of operation
+        prop_HVAC_result = get_temperatures(locator, prop_hvac).set_index('Name')
+        gv.log('done')
+
+        gv.log("calculating thermal properties")
+        prop_rc_model = self.calc_prop_rc_model(prop_occupancy, prop_architecture, prop_thermal, prop_geometry,
+                                                prop_HVAC_result, surface_properties, gv)
+        gv.log("done")
+
+        gv.log("creating windows")
+        df_windows = simple_window_generator.create_windows(surface_properties, prop_architecture)
+        gv.log("done")
+
+        # save resulting data
+        self._prop_geometry = prop_geometry
+        self._prop_architecture = prop_architecture
+        self._prop_occupancy = prop_occupancy
+        self._prop_HVAC_result = prop_HVAC_result
+        self._prop_RC_model = prop_rc_model
+        self._prop_comfort = prop_comfort
+        self._prop_internal_loads = prop_internal_loads
+        self._prop_age = prop_age
+        self._solar = solar
+        self._prop_windows = df_windows
+
+    def __len__(self):
+        return len(self.list_building_names())
+
+    def list_building_names(self):
+        """get list of all building names"""
+        return self._prop_RC_model.index
+
+    def list_uses(self):
+        """get list of all uses (occupancy types)"""
+        return list(self._prop_occupancy.drop('PFloor', axis=1).columns)
+
+    def get_prop_geometry(self, name_building):
+        """get geometry of a building by name"""
+        return self._prop_geometry.ix[name_building]
+
+    def get_prop_architecture(self, name_building):
+        """get the architecture properties of a building by name"""
+        return self._prop_architecture.ix[name_building]
+
+    def get_prop_occupancy(self, name_building):
+        """get the occupancy properties of a building by name"""
+        return self._prop_occupancy.ix[name_building]
+
+    def get_prop_hvac(self, name_building):
+        """get HVAC properties of a building by name"""
+        return self._prop_HVAC_result.ix[name_building]
+
+    def get_prop_rc_model(self, name_building):
+        """get RC-model properties of a building by name"""
+        return self._prop_RC_model.ix[name_building]
+
+    def get_prop_comfort(self, name_building):
+        """get comfort properties of a building by name"""
+        return self._prop_comfort.ix[name_building]
+
+    def get_prop_internal_loads(self, name_building):
+        """get internal loads properties of a building by name"""
+        return self._prop_internal_loads.ix[name_building]
+
+    def get_prop_age(self, name_building):
+        """get age properties of a building by name"""
+        return self._prop_age.ix[name_building]
+
+    def get_solar(self, name_building):
+        """get solar properties of a building by name"""
+        return self._solar.ix[name_building]
+
+    def get_prop_windows(self, name_building):
+        """get windows and their properties of a building by name"""
+        return self._prop_windows.loc[self._prop_windows['name_building'] == name_building].to_dict('list')
+
+    def calc_prop_rc_model(self, occupancy, architecture, thermal_properties, geometry, hvac_temperatures, surface_properties,
+                           gv):
+        """
+        Return the RC model properties for all buildings. The RC model used is described in ISO 13790:2008, Annex C (Full
+        set of equations for simple hourly method).
+
+
+        PARAMETERS
+        ----------
+
+        :param occupancy: The contents of the `occupancy.shp` file, indexed by building name. Each column is the name of an
+            occupancy type (GYM, HOSPITAL, HOTEL, INDUSTRIAL, MULTI_RES, OFFICE, PARKING, etc.) except for the
+            "PFloor" column which is a fraction of heated floor area.
+            The occupancy types must add up to 1.0.
+        :type occupancy: GeoDataFrame
+
+        :param architecture: The contents of the `architecture.shp` file, indexed by building name. It contains the
+            following fields: Occ_m2p, f_cros, n50, type_shade, win_op, win_wall. Only `win_wall` (window to wall ratio) is
+            used.
+        :type architecture: GeoDataFrame
+
+        :param thermal_properties: The contents of the `thermal_properties.shp` file, indexed by building name. It
+            contains the following fields: Es, Hs, U_base, U_roof, U_wall, U_win, th_mass.
+            - Es: fraction of gross floor area that has electricity {0 <= Es <= 1}
+            - Hs: fraction of gross floor area that is heated/cooled {0 <= Hs <= 1}
+            - th_mass: type of building construction {T1: light, T2: medium, T3: heavy}
+        :type thermal_properties: GeoDataFrame
+
+        :param geometry: The contents of the `zone.shp` file indexed by building name - the list of buildings, their floor
+            counts, heights etc.
+            Includes additional fields "footprint" and "perimeter" as calculated in `read_building_properties`.
+        :type geometry: GeoDataFrame
+
+        :param hvac_temperatures: The return value of `get_temperatures`.
+        :type hvac_temperatures: DataFrame
+
+        :param surface_properties: The contents of the `properties_surfaces.csv` file generated by the radiation script.
+            It contains the fields Name, Freeheight, FactorShade, height_ag and Shape_Leng.
+            This data is used to calculate the wall and window areas
+        :type surface_properties: DataFrame
+
+        :param gv: An instance of the GlobalVariables context.
+        :type gv: GlobalVariables
+
+
+        RETURNS
+        -------
+
+        :returns: RC model properties per building
+        :rtype: DataFrame
+
+        Sample result data:
+        Awall_all    1.131753e+03   (total wall surface exposed to outside conditions in [m2])
+        Atot         4.564827e+03   (total area of the building envelope in [m2], the roof is considered to be flat)
+        Aw           4.527014e+02   (area of windows in [m2])
+        Am           6.947967e+03   (effective mass area in [m2])
+        Aef          2.171240e+03   (floor area with electricity in [m2])
+        Af           2.171240e+03   (conditioned floor area (heated/cooled) in [m2])
+        Cm           6.513719e+08   (internal heat capacity in [J/K])
+        Htr_is       1.574865e+04   FIXME: some heat transfer coefficient
+        Htr_em       5.829963e+02   FIXME: some heat transfer coefficient
+        Htr_ms       6.322650e+04   FIXME: some heat transfer coefficient
+        Htr_op       5.776698e+02   FIXME: some heat transfer coefficient
+        Hg           2.857637e+02   (steady-state Thermal transmission coefficient to the ground in [W/K])
+        HD           2.919060e+02   (direct thermal transmission coefficient to the external environment in [W/K])
+        Htr_w        1.403374e+03   (thermal transmission coefficient for windows and glazing in [W/K])
+        GFA_m2       2.412489e+03   (gross floor area [m2])
+        Name: B153767, dtype: float64
+
+        FIXME: finish documenting the result data...
+        FIXME: rename Awall_all to something more sane...
+        """
+
+        # Areas above ground
+        # get the area of each wall in the buildings
+        surface_properties['Awall'] = (surface_properties['Shape_Leng'] * surface_properties['Freeheight'] *
+                                       surface_properties['FactorShade'])
+        df = pd.DataFrame({'Name': surface_properties['Name'],
+                           'Awall_all': surface_properties['Awall']}).groupby(by='Name').sum()
+
+        df = df.merge(architecture, left_index=True, right_index=True)
+        df = df.merge(occupancy, left_index=True, right_index=True)
+
+        # area of windows
+        df['Aw'] = df['Awall_all'] * df['win_wall'] * df['PFloor']
+
+        # opaque areas (PFloor represents a factor according to the amount of floors heated)
+        df['Aop_sup'] = df['Awall_all'] * df['PFloor'] - df['Aw']
+
+        # Areas below ground
+        df = df.merge(thermal_properties, left_index=True, right_index=True)
+        df = df.merge(geometry, left_index=True, right_index=True)
+        df = df.merge(hvac_temperatures, left_index=True, right_index=True)
+        df['floors'] = df['floors_bg'] + df['floors_ag']
+
+        # opague areas in [m2] below ground including floor
+        df['Aop_bel'] = df['height_bg'] * df['perimeter'] + df['footprint']
+
+        # total area of the building envelope in [m2], the roof is considered to be flat
+        df['Atot'] = df[['Aw', 'Aop_sup', 'footprint', 'Aop_bel']].sum(axis=1) + (df['footprint'] * (df['floors'] - 1))
+
+        df['GFA_m2'] = df['footprint'] * df['floors']  # gross floor area
+        df['Af'] = df['GFA_m2'] * df['Hs']  # conditioned area - areas not heated
+        df['Aef'] = df['GFA_m2'] * df['Es']  # conditioned area only those for electricity
+        df['Am'] = df['th_mass'].apply(self.lookup_effective_mass_area_factor) * df['Af']  # Effective mass area in [m2]
+
+        # Steady-state Thermal transmittance coefficients and Internal heat Capacity
+        df['Htr_w'] = df['Aw'] * df['U_win']  # Thermal transmission coefficient for windows and glazing in [W/K]
+
+        # direct thermal transmission coefficient to the external environment in [W/K]
+        df['HD'] = df['Aop_sup'] * df['U_wall'] + df['footprint'] * df['U_roof']
+
+        df['Hg'] = gv.Bf * df['Aop_bel'] * df[
+            'U_base']  # steady-state Thermal transmission coefficient to the ground. in W/K
+        df['Htr_op'] = df['Hg'] + df['HD']
+        df['Htr_ms'] = gv.hms * df['Am']  # Coupling conductance 1 in W/K
+        df['Htr_em'] = 1 / (1 / df['Htr_op'] - 1 / df['Htr_ms'])  # Coupling conductance 2 in W/K
+        df['Htr_is'] = gv.his * df['Atot']
+        df['Cm'] = df['th_mass'].apply(self.lookup_specific_heat_capacity) * df['Af']  # Internal heat capacity in J/K
+
+        fields = ['Awall_all', 'Atot', 'Aw', 'Am', 'Aef', 'Af', 'Cm', 'Htr_is', 'Htr_em', 'Htr_ms', 'Htr_op', 'Hg',
+                  'HD',
+                  'Htr_w', 'GFA_m2']
+        result = df[fields]
+        return result
+
+    def lookup_specific_heat_capacity(self, th_mass):
+        """
+        Look up the specific heat capacity in [J/K] for the building construction type. This is used for the calculation
+        of the internal heat capacity "Cm" in `get_prop_RC_model`.
+
+        `th_mass` is one of the following values:
+
+        - T1: light
+        - T2: medium (default)
+        - T3: heavy
+
+        :param th_mass: the type of building construction (origin: thermal_properties.shp)
+        :return:
+        """
+        if th_mass == 'T1':
+            return 110000.0
+        elif th_mass == 'T3':
+            return 300000.0
+        else:
+            return 165000.0
+
+    def lookup_effective_mass_area_factor(self, th_mass):
+        """
+        Look up the factor to multiply the conditioned floor area by to get the effective mass area by building construction
+        type. This is used for the calculation of the effective mass area "Am" in `get_prop_RC_model`.
+
+        `th_mass` is one of the following values:
+
+        - T1: light
+        - T2: medium (default)
+        - T3: heavy
+
+        :param th_mass: the type of building construction (origin: thermal_properties.shp)
+        :return: effective mass area factor
+        """
+        if th_mass == 'T2':
+            return 2.5
+        elif th_mass == 'T3':
+            return 3.2
+        elif th_mass == 'T1':
+            return 2.5
+        else:
+            return 2.5
+
+    def __getitem__(self, building_name):
+        """return a (read-only) BuildingPropertiesRow for the building"""
+        return BuildingPropertiesRow(geometry=self.get_prop_geometry(building_name),
+                                     architecture=self.get_prop_architecture(building_name),
+                                     occupancy=self.get_prop_occupancy(building_name),
+                                     hvac=self.get_prop_hvac(building_name),
+                                     rc_model=self.get_prop_rc_model(building_name),
+                                     comfort=self.get_prop_comfort(building_name),
+                                     internal_loads=self.get_prop_internal_loads(building_name),
+                                     age=self.get_prop_age(building_name),
+                                     solar=self.get_solar(building_name),
+                                     windows=self.get_prop_windows(building_name), gv=self.gv)
+
+
+class BuildingPropertiesRow(object):
+    """Encapsulate the data of a single row in the DataSets of BuildingProperties. This class meant to be
+    read-only."""
+
+    def __init__(self, geometry, architecture, occupancy, hvac,
+                 rc_model, comfort, internal_loads, age, solar, windows, gv):
+        """Create a new instance of BuildingPropertiesRow - meant to be called by BuildingProperties[building_name].
+        Each of the arguments is a pandas Series object representing a row in the corresponding DataFrame."""
+        self.geometry = geometry
+        self.architecture = architecture
+        self.occupancy = occupancy
+        self.hvac = hvac
+        self.rc_model = rc_model
+        self.comfort = comfort
+        self.internal_loads = internal_loads
+        self.age = age
+        self.solar = solar
+        self.windows = windows
+        self.building_systems = self._get_properties_building_systems(gv)
+
+    def _get_properties_building_systems(self, gv):
+        # TODO: Documentation
+        # Refactored from CalcThermalLoads
+
+        Ll = self.geometry.Blength
+        Lw = self.geometry.Bwidth
+        nf_ag = self.geometry.floors_ag
+        nf_bg = self.geometry.floors_bg
+        nfp = self.occupancy.PFloor
+        phi_pipes = self._calculate_pipe_transmittance_values()
+
+        # nominal temperatures
+        Ths_sup_0 = self.hvac.Tshs0_C
+        Ths_re_0 = Ths_sup_0 - self.hvac.dThs0_C
+        Tcs_sup_0 = self.hvac.Tscs0_C
+        Tcs_re_0 = Tcs_sup_0 + self.hvac.dTcs0_C
+        Tww_sup_0 = self.hvac.Tsww0_C
+        Tww_re_0 = Tww_sup_0 - self.hvac.dTww0_C  # Ground water temperature in heating(winter) season, according to norm #TODO: check norm
+        # Identification of equivalent lenghts
+        fforma = self._calc_form()  # factor form comparison real surface and rectangular
+        Lv = (2 * Ll + 0.0325 * Ll * Lw + 6) * fforma  # length vertical lines
+        if nf_ag < 2 and nf_bg < 2:  # it is assumed that building with less than a floor and less than 2 floors udnerground do not have
+            Lcww_dis = 0
+            Lvww_c = 0
+        else:
+            Lcww_dis = 2 * (Ll + 2.5 + nf_ag * nfp * gv.hf) * fforma  # length hot water piping circulation circuit
+            Lvww_c = (2 * Ll + 0.0125 * Ll * Lw) * fforma  # length piping heating system circulation circuit
+
+        Lsww_dis = 0.038 * Ll * Lw * nf_ag * nfp * gv.hf * fforma  # length hot water piping distribution circuit
+        Lvww_dis = (Ll + 0.0625 * Ll * Lw) * fforma  # length piping heating system distribution circuit
+
+        building_systems = pd.Series({'Lcww_dis': Lcww_dis,
+                                      'Lsww_dis': Lsww_dis,
+                                      'Lv': Lv,
+                                      'Lvww_c': Lvww_c,
+                                      'Lvww_dis': Lvww_dis,
+                                      'Tcs_re_0': Tcs_re_0,
+                                      'Tcs_sup_0': Tcs_sup_0,
+                                      'Ths_re_0': Ths_re_0,
+                                      'Ths_sup_0': Ths_sup_0,
+                                      'Tww_re_0': Tww_re_0,
+                                      'Tww_sup_0': Tww_sup_0,
+                                      'Y': phi_pipes,
+                                      'fforma': fforma})
+        return building_systems
+
+    def _calculate_pipe_transmittance_values(self):
+        """linear trasmissivity coefficients of piping W/(m.K)"""
+        if self.age.built >= 1995 or self.age.HVAC > 0:
+            phi_pipes = [0.2, 0.3, 0.3]
+        elif 1985 <= self.age.built < 1995 and self.age.HVAC == 0:
+            phi_pipes = [0.3, 0.4, 0.4]
+        else:
+            phi_pipes = [0.4, 0.4, 0.4]
+        return phi_pipes
+
+    def _calc_form(self):
+        factor = self.geometry.footprint / (self.geometry.Bwidth * self.geometry.Blength)
+        return factor
+
+
+def get_temperatures(locator, prop_HVAC):
+    """
+    Return temperature data per building based on the HVAC systems of the building. Uses the `emission_systems.xls`
+    file to look up the temperatures.
+
+    PARAMETERS
+    ----------
+
+    :param locator:
+    :type locator: LocatorDecorator
+
+    :param prop_HVAC: HVAC properties for each building (type of cooling system, control system, domestic hot water
+                      system and heating system.
+                      The values can be looked up in the contributors manual:
+                      https://architecture-building-systems.gitbooks.io/cea-toolbox-for-arcgis-manual/content/building_properties.html#mechanical-systems
+    :type prop_HVAC: GeoDataFrame
+
+    Sample data (first 5 rows):
+                 Name type_cs type_ctrl type_dhw type_hs
+    0     B154862      T0        T1       T1      T1
+    1     B153604      T0        T1       T1      T1
+    2     B153831      T0        T1       T1      T1
+    3  B302022960      T0        T0       T0      T0
+    4  B302034063      T0        T0       T0      T0
+
+
+    RETURNS
+    -------
+
+    :returns: A DataFrame containing temperature data for each building in the scenario. More information can be
+              found in the contributors manual:
+              https://architecture-building-systems.gitbooks.io/cea-toolbox-for-arcgis-manual/content/delivery_technologies.html
+    :rtype: DataFrame
+
+    Each row contains the following fields:
+    Name          B154862   (building name)
+    type_hs            T1   (copied from input)
+    type_cs            T0   (copied from input)
+    type_dhw           T1   (copied from input)
+    type_ctrl          T1   (copied from input)
+    Tshs0_C            90   (heating system supply temperature at nominal conditions [C])
+    dThs0_C            20   (delta of heating system temperature at nominal conditions [C])
+    Qhsmax_Wm2        500   (maximum heating system power capacity per unit of gross built area [W/m2])
+    Tscs0_C             0   (cooling system supply temperature at nominal conditions [C])
+    dTcs0_C             0   (delta of cooling system temperature at nominal conditions [C])
+    Qcsmax_Wm2          0   (maximum cooling system power capacity per unit of gross built area [W/m2])
+    Tsww0_C            60   (dhw system supply temperature at nominal conditions [C])
+    dTww0_C            50   (delta of dwh system temperature at nominal conditions [C])
+    Qwwmax_Wm2        500   (maximum dwh system power capacity per unit of gross built area [W/m2])
+    Name: 0, dtype: object
+
+    INPUT / OUTPUT FILES
+    --------------------
+
+    - get_technical_emission_systems: cea\db\CH\Systems\emission_systems.xls
+    """
+    prop_emission_heating = pd.read_excel(locator.get_technical_emission_systems(), 'heating')
+    prop_emission_cooling = pd.read_excel(locator.get_technical_emission_systems(), 'cooling')
+    prop_emission_dhw = pd.read_excel(locator.get_technical_emission_systems(), 'dhw')
+
+    df = prop_HVAC.merge(prop_emission_heating, left_on='type_hs', right_on='code')
+    df2 = prop_HVAC.merge(prop_emission_cooling, left_on='type_cs', right_on='code')
+    df3 = prop_HVAC.merge(prop_emission_dhw, left_on='type_dhw', right_on='code')
+
+    fields = ['Name', 'type_hs', 'type_cs', 'type_dhw', 'type_ctrl', 'Tshs0_C', 'dThs0_C', 'Qhsmax_Wm2']
+    fields2 = ['Name', 'Tscs0_C', 'dTcs0_C', 'Qcsmax_Wm2']
+    fields3 = ['Name', 'Tsww0_C', 'dTww0_C', 'Qwwmax_Wm2']
+
+    result = df[fields].merge(df2[fields2], on='Name').merge(df3[fields3], on='Name')
+    return result
